sessions-memory-demo/demo-agent/pyproject.toml
[project]
name = "sessions-memory-demo"
version = "0.1.0"
description = "ADK Sessions and Memory Bank demo for Agent Platform"
authors = [
    {name = "Carlos Cabral"},
]
dependencies = [
    "google-adk>=1.32.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.149.0",
    "protobuf>=6.31.1,<7.0.0",
    "pyopenssl>=26.0.0",
]
requires-python = ">=3.10,<3.14"


[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[project.optional-dependencies]
eval = [
    "google-adk[eval]>=1.32.0,<2.0.0",
]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ty]
[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
exclude = [".venv/**"]

[tool.ty.rules]
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
possibly-missing-attribute = "ignore"
not-subscriptable = "ignore"
deprecated = "ignore"

[tool.codespell]
ignore-words-list = "rouge"
skip = "uv.lock,.venv,**/package-lock.json"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.agents-cli]
deployment_target = "agent_runtime"
name = "sessions-memory-demo"
description = "Customer support agent with sessions and memory"
base_template = "adk"
agent_directory = "app"

generated_at = "2026-05-04T00:00:00+00:00"
acli_version = "0.0.1a1"

[tool.agents-cli.create_params]
deployment_target = "agent_runtime"
session_type = "none"
cicd_runner = "skip"
include_data_ingestion = false
datastore = "none"
